New expansion `Orders` type (#11)
Adds a new type to represent a pair of valid expansion orders.

This replaces the `[u8; 2]` argument in: `set_global_expansion_orders`, `set_expansion_on_elems`, `set_expansions_with_filter`, & `set_expansion_orders`.

This way, erroneous expansion orders can be caught when the `Orders` is constructed rather than during one of the execution of one of the above methods. 

The paper and documentation are updated to reflect this change.